CatalogProject/requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2
django-filter==21.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-seed==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==13.3.3
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
oauthlib==3.2.0
prompt-toolkit==3.0.28
psycopg==3.0.10
psycopg-binary==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toposort==1.7
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
